requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
minio
python-dotenv
asyncpg
python-multipart